requirements.txt
opencv-python==4.5.4.58
PyAutoGUI==0.9.53
pyserial==3.5